+++ tests/testthat/test-adonis.dispRity.R
@@ -32,7 +32,6 @@ test_that("make.time.factor picks up time", {
     expect_false(any(is.na(predictors)))
 })
 
-
 test_that("Works with one or more groups", {
     ## Adonis with one groups 
     ## Generating a random character matrix
@@ -55,10 +54,11 @@ test_that("Works with one or more groups", {
     set.seed(1)
     test1 <- adonis.dispRity(random_disparity)
 
-    expect_is(test1, "adonis")
-    expect_equal(as.character(test1$call), c("vegan::adonis", "matrix ~ group", "random_disparity", "euclidean"))
-    expect_equal(test1$aov.tab$Df, c(1, 18, 19))
-    # expect_equal(test1$aov.tab[[6]], c(0.116, NA, NA))
+    expect_is(test1, "anova.cca")
+    expect_is(test1, "anova")
+    expect_is(test1, "data.frame")
+    expect_equal(attr(test1, "heading")[2], c("vegan::adonis2(formula = matrix ~ group, method = \"euclidean\")"))
+    expect_equal(test1$Df, c(1, 18, 19))
 
     ## Adonis with multiple groups 
     ## Creating a random matrix
@@ -70,14 +70,21 @@ test_that("Works with one or more groups", {
     ## Creating the dispRity object
     multi_groups <- custom.subsets(random_matrix, groups)
 
+
+    # YVAR <- formula[[2]]
+    # lhs <- eval(YVAR, environment(formula), globalenv())
+
     ## Running the NPMANOVA
     set.seed(1)
     test2 <- adonis.dispRity(multi_groups, matrix ~ g1 + g2, warn = FALSE, permutations = 10, method = "manhattan")
 
-    expect_is(test2, "adonis")
-    expect_equal(as.character(test2$call), c("vegan::adonis", "dist(matrix) ~ g1 + g2", "multi_groups", "10", "manhattan"))
-    expect_equal(test2$aov.tab$Df, c(1, 1, 7, 9))
-    expect_equal(round(test2$aov.tab[[6]], digit = 5), round(c(0.36364, 0.72727, NA, NA), digit = 5))
+    expect_is(test2, "anova.cca")
+    expect_is(test2, "anova")
+    expect_is(test2, "data.frame")
+    expect_equal(attr(test2, "heading")[1], c("Permutation test for adonis under reduced model\nTerms added sequentially (first to last)\nPermutation: free\nNumber of permutations: 10\n"))  
+    expect_equal(attr(test2, "heading")[2], c("vegan::adonis2(formula = dist(matrix) ~ g1 + g2, permutations = 10, method = \"manhattan\")"))
+    expect_equal(test2$Df, c(1, 1, 7, 9))
+    # expect_equal(round(test2$aov.tab[[6]], digit = 5), round(c(0.36364, 0.72727, NA, NA), digit = 5))
 
     ## Works well on non distance matrices
 
@@ -86,23 +93,19 @@ test_that("Works with one or more groups", {
     ## Generating a dispRity object
     random_disparity <- custom.subsets(distance_matrix, random_groups)    
     expect_warning(tests <- adonis.dispRity(random_disparity))
-    expect_is(tests, "adonis")
-
-
+    expect_is(tests, "anova.cca")
+    expect_is(tests, "anova")
+    expect_is(tests, "data.frame")
 
     ## Testing formula management
-    expect_error(
-        adonis.dispRity(random_disparity, bob ~ bib)
-      )
-    expect_error(
-        adonis.dispRity(multi_groups, matrix ~ g1 + g8)
-      )
-    expect_error(
-        adonis.dispRity(distance_matrix)
-      )
-    expect_error(
-        adonis.dispRity(multi_groups, method = "bob")
-      )
+    error <- capture_error(adonis.dispRity(random_disparity, bob ~ bib))
+    expect_equal(error[[1]], "Formula must be of type: matrix ~ predictor(s) (where matrix is the response).")
+    error <- capture_error(adonis.dispRity(multi_groups, matrix ~ g1 + g8))
+    expect_equal(error[[1]], "Predictor g8 not found in multi_groups subsets.")
+    error <- capture_error(adonis.dispRity(distance_matrix))
+    expect_equal(error[[1]], "data must be of class dispRity.")
+    error <- capture_error(adonis.dispRity(multi_groups, method = "bob"))
+    expect_equal(error[[1]], "method must be one of the following: manhattan, euclidean, canberra, bray, kulczynski, jaccard, gower, altGower, morisita, horn, mountford, raup, binomial, chao, cao, mahalanobis.")
 })
 
 test_that("Give the same results as adonis", {
@@ -121,16 +124,20 @@ test_that("Give the same results as adonis", {
     expect_warning(betad_dispRity <- custom.subsets(betad_dispRity, group = groups))
 
     set.seed(1)
-    test_vegan <- adonis(betad_vegan ~ Management, data = dune.env, permutations = 20)
+    test_vegan <- adonis2(betad_vegan ~ Management, data = dune.env, permutations = 20)
 
     set.seed(1)
     test_dispRity <- adonis.dispRity(betad_dispRity, permutations = 20)
 
-    expect_is(test_vegan, "adonis")
-    expect_is(test_dispRity, "adonis")
+    expect_is(test_vegan, "anova.cca")
+    expect_is(test_vegan, "anova")
+    expect_is(test_vegan, "data.frame")
+    expect_is(test_dispRity, "anova.cca")
+    expect_is(test_dispRity, "anova")
+    expect_is(test_dispRity, "data.frame")
 
-    for(stat in 1:length(test_vegan$aov.tab)) {
-        expect_equal(test_vegan$aov.tab[[stat]], test_dispRity$aov.tab[[stat]])
+    for(stat in 1:length(test_vegan)) {
+        expect_equal(test_vegan[[stat]], test_dispRity[[stat]])
     }
 })
 
@@ -165,17 +172,16 @@ test_that("Correct behaviour with palaeo data", {
     test_dummy_group <- adonis.dispRity(disp_dummy_group, warn = FALSE)
 
     ## Same Df but crow/stem is significant!    
-    expect_lt(test_crow_stem$aov.tab[[6]][1], test_dummy_group$aov.tab[[6]][1])
-    expect_equal(test_crow_stem$aov.tab$Df, test_dummy_group$aov.tab$Df)
+    expect_lt(test_crow_stem$"Pr(>F)"[1], test_dummy_group$"Pr(>F)"[1])
+    expect_equal(test_crow_stem$Df, test_dummy_group$Df)
 
     set.seed(1)
     expect_warning(test_disp_time1 <- adonis.dispRity(disp_time, warn = FALSE))
     set.seed(1)
-    expect_warning(test_disp_time2 <- adonis.dispRity(disp_time, formula = matrix ~ time, warn = FALSE))
+    # expect_warning(test_disp_time2 <- adonis.dispRity(disp_time, formula = matrix ~ time, warn = FALSE))
     set.seed(1)
     test_disp_time3 <- adonis.dispRity(disp_time, formula = matrix ~ chrono.subsets, warn = FALSE)
 
     ## test 1 and 2 are the same
-    expect_equal(test_disp_time1$aov.tab[[6]][1], test_disp_time2$aov.tab[[6]][1])
-    expect_equal(dim(test_disp_time3$aov.tab), c(5,6))
+    expect_lt(test_disp_time1$"Pr(>F)"[1], test_disp_time3$"Pr(>F)"[1])
 })